Make centos-7 provisioning friendly to external contributors
These changes are part of an effort to make building and testing GoCD
easier outside of build.gocd and TW infrastructure. But at the same time
consitent with tools used by gocd agents running in build.gocd.

This change adds an optional argument to provision-centos.sh: --contrib
which alters behavior of the provisioning to skip several configurations
and tools which are not needed by a github contributor. When argument is
skipped, script behaves as it did before.