Skip `repoview` in CentOS 8 builds altogether.
While repoview and python2-rpmUtils RPMs do exist for CentOS/RHEL 8, none work in
practice (i.e., at runtime) due to the lack of python2 RPM bindings on the platform.

If this is _really_ needed, the easiest thing to do would be to port the script to
python3, or reimplement EVR comparison without the dependency on RPM bindings, which
appears to be the only function used from the rpmUtils library.